Hyprland

+++ .config/waybar/launch.sh
@@ -1,15 +1,26 @@
-#!/bin/bash
-# Credit to https://github.com/Alexays/Waybar/issues/961#issuecomment-753533975
+#!/usr/bin/env sh
+# vim: set ft=sh sw=4 et :
 
-CONFIG_FILES="$HOME/.config/waybar/config.jsonc $HOME/.config/waybar/style.css"
+CONFIG="$HOME/.config/waybar/config.jsonc"
+STYLE="$HOME/.config/waybar/style.css"
 
 unset GDK_BACKEND
 
+killall -q waybar
 
-trap "killall waybar" EXIT
+# Wait until the processes have been shut down
+while pgrep -u $UID -x waybar >/dev/null; do sleep 1; done
 
-while true; do
-    waybar &
-    inotifywait -e create,modify $CONFIG_FILES
-    killall waybar
-done
+mode="${1:-run}"
+cmd="waybar --config $CONFIG --style $STYLE"
+
+if [[ "$mode" = "dev" ]]; then
+    while true; do
+        WATCH_FILES="$CONFIG $STYLE"
+        $cmd --log-level=debug &
+        inotifywait -e create,modify $WATCH_FILES
+        killall waybar
+    done
+elif [[ "$mode" = "run" ]]; then
+    $cmd --log-level=warning
+fi
